requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
scikit-learn
umap-learn